\name{estDLVM-package}
\alias{estDLVM-package}
\alias{estDLVM}
\docType{package}
\title{
\packageTitle{estDLVM}
}
\description{
\packageDescription{estDLVM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{estDLVM}
\packageIndices{estDLVM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{estDLVM}

Maintainer: \packageMaintainer{estDLVM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
