\name{estDLVM Package}
\alias{estDLVM Package}
\docType{package}
\title{Overview of the package estDLVM}

\description{\packageDescription{estDLVM}}

\details{
The main functions of the package are listed hereafter:
\packageIndices{estDLVM}
}

\author{
Luca Brusa [aut, cre], Francesco Bartolucci [aut], Fulvia Pennoni [aut]

\emph{Maintainer}: Luca Brusa (luca.brusa@unimib.it)
}

\references{
Brusa, L., Bartolucci, F., Pennoni, F. (2022). Tempered expectation-maximization algorithm for the estimation of discrete latent variable models. \emph{Computational Statistics}, \bold{38}, 1--34.

Bartolucci, F., Farcomeni, A., Pennoni, F. (2013). \emph{Latent Markov Models for Longitudinal Data}, Chapman and Hall/CRC press.
}

\seealso{\code{\link{est_LC}}, \code{\link{est_HMcat}}, \code{\link{est_HMcont}}, \code{\link{est_SB}}}
